fix(sqs): route queue URL path requests to SQS handler
