+++ .gitignore
@@ -18,3 +18,4 @@ COPILOT.md
 .claude
 
 posts
+local
